+++ src/containerConfig.ts
@@ -1,4 +1,5 @@
 import { readFileSync } from 'fs';
+import { join } from 'path';
 import { container } from 'tsyringe';
 import config, { IConfig } from 'config';
 import { Probe } from '@map-colonies/mc-probe';
@@ -9,11 +10,15 @@ import { ILogger } from './common/interfaces';
 
 function registerExternalValues(): void {
   const loggerConfig = config.get<ILoggerConfig>('logger');
+  const mountDir = config.get<string>('mountDir');
+  const watchDir = config.get<string>('watcher.watchDirectory');
+  const relativeWatchDirPath = join(mountDir, watchDir);
   const packageContent = readFileSync('./package.json', 'utf8');
   const service = JSON.parse(packageContent) as IServiceConfig;
   const logger = new MCLogger(loggerConfig, service);
   container.register<IConfig>(Services.CONFIG, { useValue: config });
   container.register<ILogger>(Services.LOGGER, { useValue: logger });
+  container.register<string>(Services.WATCHER_CONFIG, { useValue: relativeWatchDirPath });
   container.register<Probe>(Probe, { useValue: getProbe(container) });
 }