+++ tests/mockData/ingestionParams.ts
@@ -18,10 +18,11 @@ export const ingestionParams = {
   ],
   metadata: {
     resolution: 0.0000018519,
+    maxResolutionMeter: 0.2,
     accuracyCE90: 1,
     classification: '4',
     creationDate: undefined,
-    sensorType: ['OTHER'],
+    sensorType: ['UNDEFINED'],
     rms: undefined,
     scale: undefined,
     description: 'description',
@@ -45,7 +46,7 @@ export const ingestionParams = {
     updateDate: '2019-04-06T00:00:00.000Z',
     sourceDateEnd: '2019-04-06T00:00:00.000Z',
     sourceDateStart: '2019-04-06T00:00:00.000Z',
-    region: undefined,
+    region: 'c1,c2',
     ingestionDate: undefined,
     srsId: undefined,
     srsName: undefined,
@@ -65,6 +66,8 @@ export const ingestionParams = {
             Rms: null,
             SensorType: 'OTHER',
             Scale: null,
+            Countries: 'c1,c2',
+            Cities: 'p1,p2,p3',
           },
           geometry: {
             type: 'Polygon',
@@ -82,5 +85,7 @@ export const ingestionParams = {
       ],
       bbox: [34.8444933529867, 32.0425133190045, 34.8851821360093, 32.0786811260038],
     },
+    includedInBests: undefined,
+    productBoundingBox: undefined,
   },
 };