+++ tests/mockData/layerMetadata.ts
@@ -2,12 +2,13 @@
 /* eslint-disable @typescript-eslint/no-magic-numbers */
 export const metadata = {
   resolution: 0.0000018519,
+  maxResolutionMeter: 0.2,
   accuracyCE90: 1,
-  sensorType: ['OTHER'],
+  sensorType: ['UNDEFINED'],
   rms: undefined,
   scale: undefined,
   producerName: undefined,
-  region: undefined,
+  region: 'c1,c2',
   classification: '4',
   creationDate: undefined,
   ingestionDate: undefined,
@@ -49,6 +50,8 @@ export const metadata = {
           Rms: null,
           SensorType: 'OTHER',
           Scale: null,
+          Countries: 'c1,c2',
+          Cities: 'p1,p2,p3',
         },
         geometry: {
           type: 'Polygon',
@@ -66,4 +69,6 @@ export const metadata = {
     ],
     bbox: [34.8444933529867, 32.0425133190045, 34.8851821360093, 32.0786811260038],
   },
+  includedInBests: undefined,
+  productBoundingBox: undefined,
 };