+++ tests/unit/layerCreator/models/classifier.spec.ts
@@ -30,7 +30,7 @@ describe('classifier', () => {
 
   describe('getClassification', () => {
     it('high res fully in bounds returns configured classification', async () => {
-      const res = 0.08;
+      const res = 0.008;
       const polygon = [
         [
           [0.01, 0.01],
@@ -49,7 +49,7 @@ describe('classifier', () => {
     });
 
     it('med res fully in bounds returns configured classification', async () => {
-      const res = 0.4;
+      const res = 0.04;
       const polygon = [
         [
           [0, 0],
@@ -68,7 +68,7 @@ describe('classifier', () => {
     });
 
     it('med res fully out of bounds returns configured classification', async () => {
-      const res = 0.4;
+      const res = 0.04;
       const polygon = [
         [
           [2, 2],
@@ -87,7 +87,7 @@ describe('classifier', () => {
     });
 
     it('med res partial cover of inner area (more then configured) returns configured classification', async () => {
-      const res = 0.4;
+      const res = 0.04;
       const polygon = [
         [
           [-0.1, -0.1],
@@ -106,7 +106,7 @@ describe('classifier', () => {
     });
 
     it('med res partial covered (less then configured) returns configured classification', async () => {
-      const res = 0.4;
+      const res = 0.04;
       const polygon = [
         [
           [-0.1, -0.1],
@@ -125,7 +125,7 @@ describe('classifier', () => {
     });
 
     it('low res in bounds returns default classification', async () => {
-      const res = 0.5;
+      const res = 0.05;
       const polygon = [
         [
           [0, 0],