+++ tests/unit/trigger/models/trigger.spec.ts
@@ -4,21 +4,23 @@ import { Trigger } from '../../../../src/layerCreator/models/trigger';
 import { HistoryStatus } from '../../../../src/layerCreator/historyStatus';
 import { overseerClientMock, ingestDiscreteLayerMock } from '../../../mocks/clients/overseerClient';
 import { agentDbClientMock, updateDiscreteStatusMock, getDiscreteStatusMock } from '../../../mocks/clients/agentDbClient';
-import {
-  validateLayerFilesExistsMock,
-  validateShpFilesExistsMock,
-  filesManagerMock,
-  fileExistsMock,
-  readAllLinesMock,
-} from '../../../mocks/filesManager';
+import { filesManagerMock, fileExistsMock, readAllLinesMock, directoryExistsMock } from '../../../mocks/filesManager';
 import { parseMock, shpParserMock } from '../../../mocks/shpParser';
 import { parseFilesShpJsonMock, mapMock, metadataMapperMock } from '../../../mocks/metadataMapperMock';
 import { lockMock, isQueueEmptyMock } from '../../../mocks/limitingLock';
 import { configMock, getMock } from '../../../mocks/config';
-import { fileMapperMock, stripSubDirsMock, getFilePathMock } from '../../../mocks/fileMapper';
+import {
+  fileMapperMock,
+  getFilePathMock,
+  getRootDirMock,
+  findFilesRelativePathsMock,
+  getFileFullPathMock,
+  cleanRelativePathMock,
+} from '../../../mocks/fileMapper';
 import { tfw } from '../../../mockData/tfw';
 import { metadata } from '../../../mockData/layerMetadata';
 import { ingestionParams } from '../../../mockData/ingestionParams';
+import { NotFoundError } from '../../../../src/common/exceptions/http/notFoundError';
 
 const expectedMetadata = loadTestMetadata();
 const expectedParams = loadTestIngestionParams();
@@ -42,6 +44,7 @@ const fileList = [
   'X1828_Y1650.Tiff',
   'X1828_Y1651.Tiff',
 ];
+const shpFiles = ['Files.shp', 'Files.dbf', 'Product.shp', 'Product.dbf', 'ShapeMetadata.shp', 'ShapeMetadata.dbf'];
 const triggeredHistoryStatus = { ...baseHistoryStatus, status: HistoryStatus.TRIGGERED };
 const failedHistoryStatus = { ...baseHistoryStatus, status: HistoryStatus.FAILED };
 
@@ -51,9 +54,9 @@ describe('trigger', () => {
   beforeEach(() => {
     configData['mountDir'] = '/mountDir';
     getMock.mockImplementation((key: string) => configData[key]);
-    stripSubDirsMock.mockImplementation((dir: string) => dir);
     getFilePathMock.mockImplementation((file: string, extension: string) => `${file}.${extension}`);
     readAllLinesMock.mockResolvedValue(tfw);
+    directoryExistsMock.mockReturnValue(true);
   });
 
   afterEach(function () {
@@ -70,11 +73,13 @@ describe('trigger', () => {
       ingestDiscreteLayerMock.mockResolvedValue({});
       updateDiscreteStatusMock.mockResolvedValue({});
       parseMock.mockResolvedValue({});
-      validateLayerFilesExistsMock.mockResolvedValue(true);
-      validateShpFilesExistsMock.mockResolvedValue(true);
       parseFilesShpJsonMock.mockReturnValue(fileList);
       mapMock.mockReturnValue(expectedMetadata);
       fileExistsMock.mockResolvedValue(true);
+      getRootDirMock.mockReturnValue('/layerSources/test');
+      findFilesRelativePathsMock.mockResolvedValueOnce(shpFiles).mockResolvedValueOnce(fileList);
+      getFileFullPathMock.mockResolvedValueOnce('file.tfw');
+      cleanRelativePathMock.mockReturnValueOnce('test');
 
       const trigger = new Trigger(
         shpParserMock,
@@ -155,8 +160,6 @@ describe('trigger', () => {
     it('manual trigger will not skip already triggered directory', async function () {
       // set mock values
       getDiscreteStatusMock.mockResolvedValue(triggeredHistoryStatus);
-      validateLayerFilesExistsMock.mockResolvedValue(true);
-      validateShpFilesExistsMock.mockResolvedValue(true);
       axiosMock.post.mockResolvedValue({});
       ingestDiscreteLayerMock.mockResolvedValue({});
       updateDiscreteStatusMock.mockResolvedValue({});
@@ -164,6 +167,9 @@ describe('trigger', () => {
       parseFilesShpJsonMock.mockReturnValue(['file.tiff']);
       mapMock.mockReturnValue(expectedMetadata);
       fileExistsMock.mockResolvedValue(true);
+      findFilesRelativePathsMock.mockResolvedValueOnce(shpFiles).mockResolvedValueOnce(['file.tiff']);
+      getFileFullPathMock.mockResolvedValueOnce('file.tfw');
+      getRootDirMock.mockReturnValue('/mountDir/test');
 
       const trigger = new Trigger(
         shpParserMock,
@@ -178,7 +184,7 @@ describe('trigger', () => {
       );
 
       // action
-      await trigger.trigger('/mountDir/test', true);
+      await trigger.trigger('test', true);
 
       // expectation
       expect(getDiscreteStatusMock).toHaveBeenCalledTimes(1);
@@ -189,8 +195,6 @@ describe('trigger', () => {
     it('manual trigger will not skip already failed directory', async function () {
       // set mock values
       getDiscreteStatusMock.mockResolvedValue(failedHistoryStatus);
-      validateLayerFilesExistsMock.mockResolvedValue(true);
-      validateShpFilesExistsMock.mockResolvedValue(true);
       axiosMock.post.mockResolvedValue({});
       ingestDiscreteLayerMock.mockResolvedValue({});
       updateDiscreteStatusMock.mockResolvedValue({});
@@ -198,6 +202,10 @@ describe('trigger', () => {
       parseFilesShpJsonMock.mockReturnValue(['file.tiff']);
       mapMock.mockReturnValue(expectedMetadata);
       fileExistsMock.mockResolvedValue(true);
+      findFilesRelativePathsMock.mockResolvedValueOnce(shpFiles).mockResolvedValueOnce(['file.tiff']);
+      getFileFullPathMock.mockResolvedValueOnce('file.tfw');
+      getRootDirMock.mockReturnValue('/mountDir/test');
+      cleanRelativePathMock.mockResolvedValue('test');
 
       const trigger = new Trigger(
         shpParserMock,
@@ -212,7 +220,7 @@ describe('trigger', () => {
       );
 
       // action
-      await trigger.trigger('/mountDir/test', true);
+      await trigger.trigger('/test', true);
 
       // expectation
       expect(getDiscreteStatusMock).toHaveBeenCalledTimes(1);
@@ -222,11 +230,12 @@ describe('trigger', () => {
 
     it('auto trigger will retry parsing invalid shp files if only one task is queued', async function () {
       // set mock values
-      validateLayerFilesExistsMock.mockResolvedValue(true);
-      validateShpFilesExistsMock.mockResolvedValue(true);
       getDiscreteStatusMock.mockResolvedValue(undefined);
       isQueueEmptyMock.mockReturnValue(true);
       parseMock.mockRejectedValue(new Error('tests'));
+      findFilesRelativePathsMock.mockResolvedValue(shpFiles);
+      getRootDirMock.mockReturnValue('/mountDir/test');
+
       configData['watcher.shpRetry'] = {
         retries: 4,
         factor: 2,
@@ -249,12 +258,50 @@ describe('trigger', () => {
 
       // action
       const action = async () => {
-        await trigger.trigger('/mountDir/test');
+        await trigger.trigger('test');
       };
 
       // expectation
       await expect(action).rejects.toThrow();
       expect(parseMock).toHaveBeenCalledTimes(5);
+    });
+
+    it('unit test: trigger will not work on invalid directory', async function () {
+      // set mock values
+      directoryExistsMock.mockReturnValue(false);
+
+      getDiscreteStatusMock.mockResolvedValue(triggeredHistoryStatus);
+      axiosMock.post.mockResolvedValue({});
+      ingestDiscreteLayerMock.mockResolvedValue({});
+      updateDiscreteStatusMock.mockResolvedValue({});
+      parseMock.mockResolvedValue({});
+      parseFilesShpJsonMock.mockReturnValue(['file.tiff']);
+      mapMock.mockReturnValue(expectedMetadata);
+      fileExistsMock.mockResolvedValue(true);
+      findFilesRelativePathsMock.mockResolvedValueOnce(shpFiles).mockResolvedValueOnce(['file.tiff']);
+      getFileFullPathMock.mockResolvedValueOnce('file.tfw');
+      getRootDirMock.mockReturnValue('/mountDir/test');
+
+      const trigger = new Trigger(
+        shpParserMock,
+        filesManagerMock,
+        metadataMapperMock,
+        overseerClientMock,
+        agentDbClientMock,
+        lockMock,
+        fileMapperMock,
+        { log: jest.fn() },
+        configMock
+      );
+
+      // action
+      const response = trigger.trigger('test', true);
+
+      // expectation
+      await expect(response).rejects.toThrow(NotFoundError);
+      expect(getDiscreteStatusMock).toHaveBeenCalledTimes(0);
+      expect(ingestDiscreteLayerMock).toHaveBeenCalledTimes(0);
+      expect(updateDiscreteStatusMock).toHaveBeenCalledTimes(0);
     });
   });
 });